Cleaned up code in server.js.
Added a controllers folder and created separate files to handle each request.
